+++ src/App.js
@@ -2,7 +2,7 @@ import { Provider } from "react-redux";
 import "./App.css";
 import HomeContent from "./pages/HomeContent";
 import Main from "./components/Main";
-import WalletContent from "./pages/WalletContent";
+import AccountContent from "./pages/AccountContent";
 import store from "./redux/store";
 import Routes from "./components/Routes";
 import ThemeContext from "./contexts/ThemeContext";
